[FRONT]error fix - user link 데이터 읽는 중
