+++ client/src/components/LoginHeader.tsx
@@ -10,6 +10,7 @@ import HeadCupIcon from '@/assets/icons/HeadCupIcon';
 
 import { useSelector } from 'react-redux';
 import { RootState } from '@/modules/store';
+import { call } from '@/utils/ApiService';
 
 type LoginProps = {
   changeNav: boolean;
@@ -59,15 +60,24 @@ const LoginHeader = ({ changeNav }:LoginProps) => {  //isUser.isLogin을 통해 
     setDropdownVariant('menu');
   };
 
-  //memberId TEST Re: 1차 시도
-  // const storedData = window.localStorage.getItem('login');
-  // const jsonID = JSON.parse(storedData);
-  // console.log('1차 json처리: ' + jsonID);
-  // const memberId = jsonID.memberId;
-  // console.log('1차 memberID : ' + memberId);
-
   const isLogin = useSelector((state:RootState) => (state.login));
   const localmemberId = isLogin.memberId;
+  console.log(localmemberId);
+
+  //임시 테스트
+  useEffect(() => {
+    const fetchUserData = async () => {
+        return call(`/users/${localmemberId}`, 'GET', null)
+        .then((res) => {
+            console.log(`USER 정보 출력:  ${res}`);
+            console.log(`USER 정보 2차 출력: ${res.nickname}`)
+        })
+        .catch((Err) => console.log('유저정보 GET 에러 발생: ' + Err));
+    };
+
+    fetchUserData();
+  }, [])
+
 
   return (
     <>
@@ -75,7 +85,7 @@ const LoginHeader = ({ changeNav }:LoginProps) => {  //isUser.isLogin을 통해 
         <ul className="flex flex-row space-x-2 items-center">
           <li className={cn(IconHover({ variant: 'hover' }))}>
             <Link to={`/users/${localmemberId}`}
-            state={{userID: localmemberId}}
+
             className="relative p-2 text-gray-500 bg-transparent"
             >
               <img
